skills/mycroft-configuration.mycroftai/requirements.txt
requests>=2.13.0
humanhash3
